+++ source/Core/ViewConfig.php
@@ -1171,39 +1171,15 @@ class ViewConfig extends \OxidEsales\Eshop\Core\Base
      */
     public function getModuleUrl($sModule, $sFile = '')
     {
-        $c = Registry::getConfig();
-        $shopUrl = null;
-        if ($this->isAdmin()) {
-            if ($c->isSsl()) {
-                // From admin and with SSL we try to use sAdminSSLURL config directive
-                $shopUrl = $c->getConfigParam('sAdminSSLURL');
-                if ($shopUrl) {
-                    // but we don't need the admin directory
-                    $adminDir = '/' . $c->getConfigParam('sAdminDir');
-                    $shopUrl = substr($shopUrl, 0, -strlen($adminDir));
-                } else {
-                    // if no sAdminSSLURL directive were defined we use sSSLShopURL config directive instead
-                    $shopUrl = $c->getConfigParam('sSSLShopURL');
-                }
-            }
-            // From admin and with no config usefull directive, we use the sShopURL directive
-            if (!$shopUrl) {
-                $shopUrl = $c->getConfigParam('sShopURL');
-            }
-        }
-        // We are either in front, or in admin with no $sShopURL defined
-        if (!$shopUrl) {
-            $shopUrl = $c->getCurrentShopUrl();
-        }
-        $shopUrl = rtrim($shopUrl, '/');
+        $config = Registry::getConfig();
 
-        $sUrl = str_replace(
-            rtrim($c->getConfigParam('sShopDir'), '/'),
-            $shopUrl,
+        $moduleUrl = str_replace(
+            rtrim($config->getConfigParam('sShopDir'), '/'),
+            rtrim($config->getCurrentShopUrl(false), '/'),
             $this->getModulePath($sModule, $sFile)
         );
 
-        return $sUrl;
+        return $moduleUrl;
     }
 
     /**